automatically backed up by learn

+++ lib/bachelor.rb
@@ -54,5 +54,5 @@ def get_average_age_for_season(data, season)
       end
     end
   end
-  return (sum_age / i).to_i.ceil
+  return (sum_age / i).to_f.ceil
 end
